+++ lib/src/cursor.dart
@@ -22,17 +22,33 @@ class SQLiteCursor extends MatrixCursor {
   SQLiteCursor.from({@required List<String> columns, @required List<List<dynamic>> rows})
     : super.from(columns: columns, rows: rows);
 
-  /// Returns the value of the requested column as a UTC instant in time.
+  /// Returns the value of the requested column as a boolean.
   ///
-  /// See: https://www.sqlite.org/lang_datefunc.html
-  DateTime getDateTime(final int columnIndex) {
+  /// See: https://www.sqlite.org/datatype3.html#boolean_datatype
+  bool getBool(final int columnIndex) {
     switch (getType(columnIndex)) {
       case Cursor.FIELD_TYPE_NULL:
         return null;
       case Cursor.FIELD_TYPE_INTEGER:
-        return DateTime.fromMillisecondsSinceEpoch(getInt(columnIndex), isUtc: true);
-      case Cursor.FIELD_TYPE_STRING:
+        return getInt(columnIndex) != 0;
+      default:
+        throw SQLiteDatatypeMismatchException();
+    }
+  }
+
+  /// Returns the value of the requested column as a UTC instant in time.
+  ///
+  /// See: https://www.sqlite.org/datatype3.html#date_and_time_datatype
+  DateTime getDateTime(final int columnIndex) {
+    switch (getType(columnIndex)) {
+      case Cursor.FIELD_TYPE_NULL:
+        return null;
+      case Cursor.FIELD_TYPE_STRING:  // ISO-8601 strings ("YYYY-MM-DD HH:MM:SS.SSS")
         return DateTime.parse(getString(columnIndex)); // TODO: default timezone?
+      case Cursor.FIELD_TYPE_INTEGER: // Unix time, seconds since 1970-01-01T00:00:00Z
+        return DateTime.fromMillisecondsSinceEpoch(getInt(columnIndex) * 1000, isUtc: true);
+      case Cursor.FIELD_TYPE_FLOAT:   // Julian day numbers
+        // TODO: implement support for Julian day numbers; for now, fall through:
       default:
         throw SQLiteDatatypeMismatchException();
     }