+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.6] - 2018-11-30
 ### Added
+- `SQLiteCursor#getBool()` method
 
 ## [0.1.5] - 2018-11-30
 ### Added